File_Handler updates

+++ src/commons/file_handler.py
@@ -16,296 +16,308 @@ predict_root = 'predict_files'
 training_root = 'training_files'
 predictor = 'predictor'
 
-# -------------------------conversion functions------------------------------------
 
 
+#--------------------------convertion functions-------------------------------------
 # csv -> parquet
-def csv2Parquet(input_path: str, output_path: str):
-    cf = pd.read_csv(input_path, index_col=False)
-    cf.to_parquet(output_path)
+def csv2Parquet(input_path:str , output_path:str):
+  if not fileExistanceCheck([input_path]):
+    print('CSV file not found')
+    return 'CSV file not found'
+  cf = pd.read_csv(input_path, index_col = False)
+  cf.to_parquet(output_path)
 
 
 # parquet -> csv
-def parquet2Csv(input_path: str, output_path: str):
-    pf = pd.read_parquet(input_path)
-    pf.to_csv(output_path, index = False)
+def parquet2Csv(input_path:str, output_path:str):
+  if not fileExistanceCheck([input_path]):
+    print('Parquet file not found')
+    return 'Parquet file not found'
+  pf = pd.read_parquet(input_path)
+  pf.to_csv(output_path, index = False)
 
 
 # csv -> json format
-def csv2Json(input_path: str, output_path: str):
-    cf = pd.read_csv(input_path)
-    cf.to_json(output_path, orient='records')
+def csv2Json(input_path:str, output_path:str):
+  if not fileExistanceCheck([input_path]):
+    print('CSV file not found')
+    return 'CSV file not found'
+  cf = pd.read_csv(input_path)
+  cf.to_json(output_path, orient='records')
 
 
 # json file -> csv
-def json2Csv(input_path: str, output_path: str):
-    jf = pd.read_json(input_path)
-    jf.to_csv(output_path, index=False)
+def json2Csv(input_path:str, output_path:str):
+  if not fileExistanceCheck([input_path]):
+    print('Json file not found')
+    return 'Json file not found'
+  jf = pd.read_json(input_path)
+  jf.to_csv(output_path, index=False)
+  print(f'File has been saved at {output_path}')
 
 
-# csv -> pickle
-def csv2Pickle(input_path: str, output_path: str):
-    cf = pd.read_csv(input_path)
-    cf.to_pickle(output_path)
 
+def parquetGenerateCsv(input_path:str):
+  if not fileExistanceCheck([input_path]):
+    print('Parquet file not found')
+    return 'Parquet file not found'
+  file_path,full_name = os.path.split(input_path)
+  file_name = full_name.split('.')[0]
+  new_log = os.path.join(file_path, file_name + '.csv')
+  print('Parquet->CSV start...')
+  parquet2Csv(input_path, new_log)
+  print('Parquet->CSV finished...')
+  return new_log
 
-# pickle -> csv
-def pickle2Csv(input_path: str, output_path: str):
-    pf = pd.read_pickle(input_path)
-    pf.to_csv(output_path)
-
-
-def parquetGenerateCsv(uuid: str, file_name: str, input_address: str):
-    filename, extension = os.path.splitext(file_name)
-    new_log = loadPredictEventLogAddress(uuid, filename) + '.csv'
-    print('Parquet->CSV start...')
-    parquet2Csv(input_address, new_log)
-    print('Parquet->CSV finished...Remove Parquet...')
-    removeFile(input_address)
-    return new_log
-
-
-# ---------------------------file loading functions-------------------------------
+#----------------------------file loading functions-------------------------------------------------------
 # loading CSV file into String format
-def csvLoadingAsString(input_path: str):
-    cf = pd.read_csv(input_path, index_col=False)
-    s = cf.to_json()
-    return s
+def csvLoadingAsString(input_path):
+  cf = pd.read_csv(input_path, index_col= False)
+  s = cf.to_json()
+  return s
 
 
 # loading CSV file into json dict
-def csvLoadingAsDict(input_path: str):
-    s = csvLoadingAsString(input_path)
-    jd = json.loads(s)
-    return jd
-
-
-def csvLoadingHead(input_path: str):
-    ch = pd.read_csv(input_path)
-    return ch.head()
+def csvLoadingAsDict(input_path:str):
+  s = csvLoadingAsString(input_path)
+  jd = json.loads(s)
+  return jd
 
 
 # loading json file into string(jsonFormat)(used inside)
-def jsonLoadingAsString(js_path: str):
-    jf = pd.read_json(js_path)
-    s = jf.to_json()
-    return s
+def jsonLoadingAsString(js_path:str):
+  jf = pd.read_json(js_path)
+  s = jf.to_json()
+  return s
 
 
 # loading json file into dict
-def jsonLoadingAsDict(js_path: str):
-    s = jsonLoadingAsString(js_path)
-    jd = json.loads(s)
-    return jd
+def jsonLoadingAsDict(js_path:str):
+  s = jsonLoadingAsString(js_path)
+  jd = json.loads(s)
+  return jd
 
 
 # loading pickle file into dict
-def pickleLoadingAsDict(pickle_path: str):
-    with open(pickle_path, 'rb') as pf:
-        _pd = pickle.load(pf)
-    return _pd
+def pickleLoadingAsDict(pickle_path:str):
+  if not os.path.exists(pickle_path):
+    print('pickle not exist')
+    return 'pickle not exist'
+
+  with open(pickle_path,'rb') as pf:
+    pd = pickle.load(pf)
+  return pd
 
 
 # load root address
-def loadPredictRoot(uuid: str,  volume_address=''):
+def loadPredictRoot(uuid:str,  additional_address = ''):
+  address = os.path.join(additional_address,predict_root,uuid)
 
-    address = os.path.join(volume_address, predict_root, uuid)
-    return address
+  if not os.path.exists(address):
+    print(f'Task {uuid} not exist')
+    return 'Root not exist'
+
+  return address
 
 
-def loadTrainingRoot(uuid: str,  volume_address=''):
-    address = os.path.join(volume_address, training_root, uuid)
+def loadTrainingRoot(uuid:str,  additional_address = ''):
+  address = os.path.join(additional_address,training_root,uuid)
 
-    return address
+  if not os.path.exists(address):
+    print(f'Task {uuid} not exist')
+    return 'Root not exist'
 
-# ----------------------------------Eventlog functions-------------------------------
+  return address
+#-----------------------------------Eventlog functions----------------------------------------------
 # save predictive monitor csv file
-# Volume address logic needs to be solved later
-def savePredictEventlog(uuid: str, file: UploadFile, volume_address = ''):
-    root_address = os.path.join(volume_address, predict_root, uuid)
-    folder = os.path.exists(root_address)
+def savePredictEventlog(uuid: str, file: UploadFile, additional_address = ''):
+  if not csvCheck(file.filename) and not parquetCheck(file.filename):
+      return 'Eventlog File not accepted'
 
-    if not folder:
-        os.makedirs(root_address)
+  root_address = os.path.join(additional_address,predict_root,uuid)
 
-    with open(os.path.join(root_address, file.filename), 'wb') as buffer:
-        shutil.copyfileobj(file.file, buffer)
+  
+  return saveFile(root_address, file)
 
 
 # load EventLog address
-def loadPredictEventLogAddress(uuid: str, file_name: str, volume_address=''):
-    root_address = os.path.join(volume_address, predict_root, uuid, file_name)
+def loadPredictEventLogAddress(uuid: str, file_name: str, additional_address = ''):
+  root_address = root_address = os.path.join(additional_address,predict_root,uuid,file_name)
+  if not os.path.exists(root_address):
+    return 'Eventlog not found'
 
-    return root_address
+  return root_address
 
 
 # save training Eventlog
-def saveTrainingEventlog(uuid: str, file: UploadFile, volume_address=''):
-    root_address = os.path.join(volume_address, training_root, uuid)
-    folder = os.path.exists(root_address)
+def saveTrainingEventlog(uuid: str, file:UploadFile, additional_address = ''):
+  if not csvCheck(file.filename) and not parquetCheck(file.filename):
+      return 'Eventlog File not accepted'
 
-    if not folder:
-        os.makedirs(root_address)
-
-    with open(os.path.join(root_address, file.filename), 'wb') as buffer:
-        shutil.copyfileobj(file.file, buffer)
+  root_address = os.path.join(additional_address,training_root,uuid)
+  
+  return saveFile(root_address, file)
 
 
-# load training event log address
-def loadTrainingEventLogAddress(uuid: str, file_name: str, volume_address=''):
-    root_address = os.path.join(volume_address, training_root, uuid, file_name)
+#load training Eventlog address
+def loadTrainingEventLogAddress(uuid:str, file_name:str, additional_address = ''):
+  root_address = root_address = os.path.join(additional_address,training_root,uuid,file_name)
 
-    return root_address
+  return root_address
 
-
-# --------------------------------Pickle functions---------------------------------------------
+#---------------------------------Pickle functions---------------------------------------------
 # save pickle dict as pickle file
-def savePredictor(uuid: str, files: List[UploadFile], volume_address=''):
-    root_address = os.path.join(volume_address, predict_root, uuid, predictor)
-    folder = os.path.exists(root_address)
+def savePredictor(uuid: str, files:List[UploadFile], additional_address = ''):
+  if files == []:
+    return 'No predictor'
 
-    if not folder:
-        os.makedirs(root_address)
+  for pfile in files:
+    if not pickleCheck(pfile.filename):
+      return 'Pickle file not accept'
 
-    for pfile in files:
-        with open(os.path.join(root_address, pfile.filename), 'wb') as buffer:
-            shutil.copyfileobj(pfile.file, buffer)
+  for pfile in files:
+    root_address = os.path.join(additional_address,predict_root,uuid,predictor)
+    saveFile(root_address, pfile)
+  
+  return 'Pickles saved'
 
 
 # load pickle file address by uuid and name
-def loadPredictorAddress(uuid: str, volume_address=''):
-    root_address = os.path.join(volume_address, predict_root, uuid, predictor)
+def loadPredictorAddress(uuid: str, additional_address = ''):
+  root_address = root_address = os.path.join(additional_address,predict_root,uuid,predictor)
 
-    return root_address
+  if not os.path.exists(root_address):
+    return 'Predictor does not exists'
 
+  return root_address
 
-# -------------------------------Schema functions-------------------------------------------------
+#--------------------------------Schema functions-------------------------------------------------
 # save Schema dict as pickle file
-def savePredictSchema(uuid: str, file: UploadFile, volume_address=''):
-    root_address = os.path.join(volume_address, predict_root, uuid)
-    folder = os.path.exists(root_address)
-
-    if not folder:
-        os.makedirs(root_address)
-
-    with open(os.path.join(root_address, file.filename), 'wb') as buffer:
-        shutil.copyfileobj(file.file, buffer)
-
-
-def saveTrainingSchema(uuid: str, file: UploadFile, volume_address=''):
-    root_address = os.path.join(volume_address, training_root, uuid)
-    folder = os.path.exists(root_address)
-
-    if not folder:
-        os.makedirs(root_address)
-
-    with open(os.path.join(root_address,file.filename), 'wb') as buffer:
-        shutil.copyfileobj(file.file, buffer)
-
+def savePredictSchema(uuid: str, file: UploadFile, additional_address = ''):
+  if not schemaCheck(file.filename):
+    return 'Schema file not accept'
+  root_address = os.path.join(additional_address,predict_root,uuid)
+  
+  return saveFile(root_address, file)
 
 # load pickle file address by uuid and name
-def loadPredictSchemaAddress(uuid: str, file_name: str, volume_address=''):
+def loadPredictSchemaAddress(uuid: str, file_name: str, additional_address = ''):
+  root_address = root_address = os.path.join(additional_address,predict_root,uuid,file_name)
+  if not os.path.exists(root_address):
+    return 'Schema not found'
 
-    root_address = os.path.join(volume_address, predict_root, uuid, file_name)
-    return root_address
+  return root_address
 
 
-def loadTrainingSchemaAddress(uuid: str, file_name: str, volume_address=''):
+def saveTrainingSchema(uuid: str, file: UploadFile, additional_address = ''):
+  if not schemaCheck(file.filename):
+    return 'Schema file not accept'
 
-    root_address = os.path.join(volume_address, training_root, uuid, file_name)
-    return root_address
+  root_address = os.path.join(additional_address,training_root,uuid)
 
+  return saveFile(root_address, file)
+
+
+def loadTrainingSchemaAddress(uuid: str, file_name: str, additional_address = ''):
+  root_address = root_address = os.path.join(additional_address,training_root,uuid,file_name)
+  if not os.path.exists(root_address):
+    return 'Schema not found'
+
+  return root_address
 
 # -----------------------------Config functions---------------------------------------------------
 def saveConfig(uuid: str, file: UploadFile, volume_address=''):
-    root_address = os.path.join(volume_address, training_root, uuid)
-    folder = os.path.exists(root_address)
+  if not schemaCheck(file.filename):
+    return 'Config file not accept'
 
-    if not folder:
-        os.makedirs(root_address)
+  root_address = os.path.join(volume_address, training_root, uuid)
 
-    with open(os.path.join(root_address, file.filename), 'wb') as buffer:
-        shutil.copyfileobj(file.file, buffer)
+  return saveFile(root_address, file)
 
 
 # load EventLog address
 def loadConfigAddress(uuid: str, file_name: str, volume_address=''):
     root_address = os.path.join(volume_address, training_root, uuid, file_name)
+    if not os.path.exists(root_address):
+      return 'Config not found'
 
     return root_address
 
+#------------------------------Result functions---------------------------------------------------
+def loadPredictResult(uuid: str,additional_address=''):
+  root_address = os.path.join(additional_address,predict_root,uuid)
+  if not os.path.exists(root_address):
+    return 'Task not exist'
 
-# -----------------------------Result functions---------------------------------------------------
-def loadPredictResult(uuid: str, volume_address=''):
-
-    root_address = os.path.join(volume_address, predict_root, uuid)
-    allFile = os.listdir(root_address)
-
-    for file in allFile:
-        if file == uuid + '-results.csv':
-            return os.path.join(root_address, file)
-
-
-def loadTrainingResult(uuid: str, volume_address=''):
-
-    root_address = os.path.join(volume_address, training_root, uuid)
-    allFile = os.listdir(root_address)
-
-    for file in allFile:
-        if file == uuid + '-results.zip':
-            return os.path.join(root_address, file)
-
-
-# -----------------------------File checking functions---------------------------------------------------
-# check file existence
-def fileExistenceCheck(files: List):
-    result = True
-
-    for file in files:
-        if not os.path.exists(file):
-            result = False
+  allFile = os.listdir(root_address)
   
-    return result
+  for file in allFile:
+    if file == uuid + '-results.csv':
+      return os.path.join(root_address,file)
+
+  return 'Result not found'
+
+
+def loadTraingingResult(uuid:str, additional_address=''):
+  root_address = os.path.join(additional_address,training_root,uuid)
+  allFile = os.listdir(root_address)
+
+  for file in allFile:
+    if file == uuid + '-results.pkl':
+      return os.path.join(root_address,file)
+
+
+#------------------------------File checking functions---------------------------------------------------
+# check file existance
+def fileExistanceCheck(files: List):
+  result = True
+
+  for file in files:
+    if not os.path.exists(file):
+      result = False
+  
+  return result
 
 
 # check the file in csv format
 def csvCheck(file: str):
-    filename, extension = os.path.splitext(file)
 
-    if extension == '.csv':
-        return True
-    else:
-        return False
+  filename,extension = os.path.splitext(file)
+
+  if extension == '.csv':
+    return True
+  else:
+    return False
 
 
 # check the  file in pickle format
 def pickleCheck(file: str):
-    filename, extension = os.path.splitext(file)
+  filename,extension = os.path.splitext(file)
   
-    if extension == '.pkl' or extension == '.pickle':
-        return True
-    else:
-        return False
+  if extension == '.pkl' or extension == '.pickle':
+    return True
+  else:
+    return False
 
 
 # check schema file in json format
 def schemaCheck(file: str):
-    filename, extension = os.path.splitext(file)
+  filename,extension = os.path.splitext(file)
   
-    if extension == '.json':
-        return True
-    else:
-        return False
+  if extension == '.json':
+    return True
+  else:
+    return False
 
 
 # check parquet file
 def parquetCheck(file: str):
-    filename, extension = os.path.splitext(file)
+  filename,extension = os.path.splitext(file)
   
-    if extension == '.parquet':
-        return True
-    else:
-        return False
+  if extension == '.parquet':
+    return True
+  else:
+    return False
 
 
 # check config file in json format
@@ -317,6 +329,7 @@ def configCheck(file: str):
     else:
         return False
 
+  
 
 # ------------------------------zip functions-----------------------------------------------------
 def zipFile(uuid: str, keep_files: bool = True, volume_address: str = ''):
@@ -350,28 +363,44 @@ def loadZip(uuid: str, volume_address=''):
 
     return zip_address
 
-
-# -------------------------------Delete functions-------------------------------------------------
-def removePredictTaskFile(uuid: str, volume_address=''):
-    rm_pass = os.path.join(volume_address, predict_root, uuid)
-    shutil.rmtree(rm_pass, onerror=lambda func, path, excinfo: print(excinfo))
-
-
-def removeTrainingTaskFile(uuid: str, volume_address=''):
-    rm_pass = os.path.join(volume_address, training_root, uuid)
-    shutil.rmtree(rm_pass, onerror=lambda func, path, excinfo: print(excinfo))
+#--------------------------------Delete functions-------------------------------------------------
+def removePredictTaskFile(uuid: str, additional_address = ''):
+  rm_pass = os.path.join(additional_address, predict_root, uuid)
+  if os.path.exists(rm_pass):
+    shutil.rmtree(rm_pass)
+    return f'Task {uuid} has been deleted'
+  else:
+    return f'Task {uuid} not found'
 
 
-def removeFile(path: str):
-    try:
-        os.remove(path)
-    except OSError as err:
-        print(err)
+def removeTrainingTaskFile(uuid: str, additional_address=''):
+  rm_pass = os.path.join(additional_address, training_root, uuid)
+  if os.path.exists(rm_pass):
+    shutil.rmtree(rm_pass)
+    return f'Task {uuid} has been deleted'
+  else:
+    return f'Task {uuid} not found'
 
 
-# -----------------------------serializing functions---------------------------------------------------
-def baseDecode(base:str):
-    return base64.decode(base)
+def removeFile(path:str):
+  if os.path.exists(path):
+    os.remove(path)
+    return f'File {path} has been deleted'
+  else:
+    return 'File not found'
+    
+#------------------------------common functions---------------------------------------------------
+def saveFile(address:str, file:UploadFile):
+
+  if not os.path.exists(address):
+    os.makedirs(address)
+
+  path = os.path.join(address,file.filename)
+
+  with open(path, 'wb') as buffer:
+    shutil.copyfileobj(file.file, buffer)
+    return path
+
 
 
 ## Docker volume test ##
